+++ apis/nucleus/src/components/listbox/ListBoxInline.jsx
@@ -80,7 +80,7 @@ function ListBoxInline({ options, layout }) {
     showLock = false,
     components,
     selectDisabled = () => false,
-    flags,
+    disablePortal = true,
   } = options;
 
   const theme = useTheme();
@@ -88,8 +88,7 @@ function ListBoxInline({ options, layout }) {
   const { translator, keyboardNavigation, themeApi, constraints } = useContext(InstanceContext);
 
   const { checkboxes = checkboxesOption } = layout || {};
-  const themeSelectionColorsEnabled = flags?.isEnabled('PS_22149_THEME_SELECTION_COLORS');
-  const styles = getStyles({ app, themeApi, theme, components, checkboxes, themeSelectionColorsEnabled });
+  const styles = getStyles({ app, themeApi, theme, components, checkboxes });
 
   const isDirectQuery = isDirectQueryEnabled({ appLayout: app?.layout });
 
@@ -252,6 +251,7 @@ function ListBoxInline({ options, layout }) {
       selectionState={selectionState}
       selections={selections}
       keyboard={keyboard}
+      disablePortal={disablePortal}
     />
   );