+++ commands/serve/CHANGELOG.md
@@ -3,6 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.9.0](https://github.com/qlik-oss/nebula.js/compare/v5.8.1...v5.9.0) (2025-01-13)
+
+### Bug Fixes
+
+- **deps:** update dependency chokidar to v4 ([#1608](https://github.com/qlik-oss/nebula.js/issues/1608)) ([4cbeba7](https://github.com/qlik-oss/nebula.js/commit/4cbeba7555c0a43a588e0141ba210ef4b61f667a))
+- **deps:** update dependency puppeteer to v23 ([#1588](https://github.com/qlik-oss/nebula.js/issues/1588)) ([417bb81](https://github.com/qlik-oss/nebula.js/commit/417bb81d25a063cf9bd70dc6a900700baf943845))
+- **deps:** update dependency webpack-cli to v6 ([#1656](https://github.com/qlik-oss/nebula.js/issues/1656)) ([6d47e5b](https://github.com/qlik-oss/nebula.js/commit/6d47e5bea30ff43e8fd03fcf9a634880d0ac3f07))
+- **deps:** update minor and patch ([#1649](https://github.com/qlik-oss/nebula.js/issues/1649)) ([cd22432](https://github.com/qlik-oss/nebula.js/commit/cd22432c37ddfffec19109eb6366adfcac32a36c))
+- **deps:** update minor and patch ([#1655](https://github.com/qlik-oss/nebula.js/issues/1655)) ([123f2ef](https://github.com/qlik-oss/nebula.js/commit/123f2ef95b5491e789015a62f6bce048d20984fb))
+- **deps:** update minor and patch ([#1660](https://github.com/qlik-oss/nebula.js/issues/1660)) ([cb43faa](https://github.com/qlik-oss/nebula.js/commit/cb43faa739a4a2e06b178dcd22c90092a36276d0))
+
+## [5.8.1](https://github.com/qlik-oss/nebula.js/compare/v5.8.0...v5.8.1) (2024-11-23)
+
+**Note:** Version bump only for package @nebula.js/cli-serve
+
+# [5.8.0](https://github.com/qlik-oss/nebula.js/compare/v5.7.1...v5.8.0) (2024-11-18)
+
+**Note:** Version bump only for package @nebula.js/cli-serve
+
+## [5.7.1](https://github.com/qlik-oss/nebula.js/compare/v5.7.0...v5.7.1) (2024-10-15)
+
+**Note:** Version bump only for package @nebula.js/cli-serve
+
+# [5.7.0](https://github.com/qlik-oss/nebula.js/compare/v5.6.0...v5.7.0) (2024-10-15)
+
+**Note:** Version bump only for package @nebula.js/cli-serve
+
+# [5.6.0](https://github.com/qlik-oss/nebula.js/compare/v5.5.0...v5.6.0) (2024-10-08)
+
+**Note:** Version bump only for package @nebula.js/cli-serve
+
+# [5.5.0](https://github.com/qlik-oss/nebula.js/compare/v5.4.2...v5.5.0) (2024-09-30)
+
+### Bug Fixes
+
+- **deps:** update dependency webpack to v5.94.0 [security] ([#1600](https://github.com/qlik-oss/nebula.js/issues/1600)) ([db8e12a](https://github.com/qlik-oss/nebula.js/commit/db8e12a016d82e176d1cffab5c6b7da2c8699d53))
+
 ## [5.4.2](https://github.com/qlik-oss/nebula.js/compare/v5.4.1...v5.4.2) (2024-08-27)
 
 **Note:** Version bump only for package @nebula.js/cli-serve